+++ lib/src/script/script_element.dart
@@ -38,8 +38,9 @@ abstract class ScriptElement extends Equatable {
   static const CONST_OP_EQUALVERIFY = 0x88;
 
   static const int CONST_OP_HASH160 = 0xa9;
+  static const int CONST_OP_CHECKSIG = 0xac;
 
-  int run(Stack<ScriptElement> stack);
+  int run(Stack<ScriptData> stack);
 
   Uint8List get toUint8List;
 }
@@ -50,7 +51,8 @@ class ScriptData extends ScriptElement {
   ScriptData(this.data);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
+    stack.push(this);
     return 1; //Do nothing
   }
 
@@ -67,10 +69,20 @@ class ScriptData extends ScriptElement {
 
   @override
   Uint8List get toUint8List => data;
+
+  @override
+  String toString() {
+    return "ScriptData(${data.toHex})";
+  }
 }
 
 class ScriptNumber extends ScriptData {
   ScriptNumber(int number) : super(number.to8Bits());
+
+  @override
+  String toString() {
+    return "ScriptNumber(${data.buffer.asByteData().getUint8(0)})";
+  }
 }
 
 abstract class ScriptOperation extends ScriptElement {
@@ -103,17 +115,22 @@ abstract class ScriptOperation extends ScriptElement {
 
   @override
   Uint8List get toUint8List => opCode.to8Bits();
+
+  @override
+  String toString() {
+    return "${runtimeType.toString()}()";
+  }
 }
 
 class ScriptOpDup extends ScriptOperation {
   ScriptOpDup() : super(ScriptElement.CONST_OP_DUP);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.isEmpty) {
       return 1;
     }
-    ScriptElement element = stack.pop();
+    ScriptData element = stack.pop();
     stack.push(element);
     stack.push(element);
     return 1;
@@ -124,7 +141,7 @@ class ScriptOpDepth extends ScriptOperation {
   ScriptOpDepth() : super(0x74);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     stack.push(ScriptNumber(stack.length));
     return 1;
   }
@@ -135,15 +152,16 @@ class ScriptOpAdd extends ScriptOperation {
   ScriptOpAdd(): super(0x93);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.length >= 2) {
-      final element1 = stack.pop() as ScriptData;
-      final element2 = stack.pop() as ScriptData;
+      final element1 = stack.pop();
+      final element2 = stack.pop();
       final value = element1.data.toUint8List.buffer.asByteData().getUint8(0) + element2.data.toUint8List.buffer.asByteData().getUint8(0);
       stack.push(ScriptNumber(value));
+      return 1;
     }
 
-    return -1;
+    return 0;
   }
 }
 
@@ -152,7 +170,7 @@ class ScriptOpDrop extends ScriptOperation {
   ScriptOpDrop() : super(0x75);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.isNotEmpty) {
       stack.pop();
     }
@@ -166,17 +184,15 @@ class ScriptOpHash256 extends ScriptOperation {
   ScriptOpHash256() : super(0xaa);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.isEmpty) {
       return 0;
     }
 
-    ScriptElement element = stack.pop();
-    if(element is ScriptData) {
-      final hashed = Hash.sha256(Hash.sha256(element.data));
-      stack.push(ScriptData(hashed));
-    }
-    return 0;
+    ScriptData element = stack.pop();
+    final hashed = Hash.sha256(Hash.sha256(element.data));
+    stack.push(ScriptData(hashed));
+    return 1;
   }
 }
 
@@ -185,19 +201,15 @@ class ScriptOpHash160 extends ScriptOperation {
   ScriptOpHash160() : super(ScriptElement.CONST_OP_HASH160);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.isEmpty) {
       return 0;
     }
 
-    ScriptElement element = stack.pop();
-    if(element is ScriptData) {
-      final hashed = Hash.hash160(element.data);
-      stack.push(ScriptData(hashed));
-      return 1;
-    }
-
-    return 0;
+    ScriptData element = stack.pop();
+    final hashed = Hash.hash160(element.data);
+    stack.push(ScriptData(hashed));
+    return 1;
   }
 
 }
@@ -207,13 +219,14 @@ class ScriptOpEqual extends ScriptOperation {
   ScriptOpEqual() : super(ScriptElement.CONST_OP_EQUAL);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.length >= 2) {
       final element1 = stack.pop();
       final element2 = stack.pop();
 
-      final result = ScriptNumber((element1 == element2) ? 1 : 0);
+      final result = ScriptNumber((element1.data.compare(element2.data)) ? 1 : 0);
       stack.push(result);
+      return 1;
     }
 
     return 0;
@@ -226,7 +239,7 @@ class ScriptOpEqualVerify extends ScriptOperation {
   ScriptOpEqualVerify() : super(ScriptElement.CONST_OP_EQUALVERIFY);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(ScriptOperation.OP_EQUAL.run(stack) > 0) {
       return ScriptOperation.OP_VERIFY.run(stack);
     }
@@ -237,14 +250,14 @@ class ScriptOpEqualVerify extends ScriptOperation {
 
 class ScriptOpCheckSig extends ScriptOperation {
 
-  ScriptOpCheckSig() : super(0xac);
+  ScriptOpCheckSig() : super(ScriptElement.CONST_OP_CHECKSIG);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.length >= 2) {
-      final pubKey = stack.pop() as ScriptData;
-      final sig = stack.pop() as ScriptData;
-      final sigHash = SignedHash.fromDER(pubKey.data, sig.data);
+      final pubKey = stack.pop();
+      final sig = stack.pop();
+      final sigHash = SignedHash.fromDER(pubKey.data, sig.data.sublist(0, sig.data.length - 1));
       final verified = sigHash.verify(PublicKey.fromSEC(EllipticCurve.secp256k1, pubKey.data));
       final result = (verified) ? 1 : 0;
       stack.push(ScriptNumber(result));
@@ -260,12 +273,11 @@ class ScriptOpVerify extends ScriptOperation {
   ScriptOpVerify() : super(ScriptElement.CONST_OP_VERIFY);
 
   @override
-  int run(Stack<ScriptElement> stack) {
+  int run(Stack<ScriptData> stack) {
     if(stack.isNotEmpty) {
       final element = stack.pop();
-      if(element is ScriptData) {
-        return element.data.buffer.asByteData().getUint8(0);
-      }
+
+      return element.data.buffer.asByteData().getUint8(0);
     }
     return 0;
   }