+++ lib/src/trx/sighash.dart
@@ -1,0 +1,10 @@
+class SigHash {
+  static const int SIGHASH_FORKID = 0x40;// 0100 0000
+  static const int SIGHASH_ALL = 0x41;   // 0100 0001
+  static const int SIGHASH_NONE = 0x42;  // 0100 0010
+  static const int SIGHASH_SINGLE = 0x43;// 0100 0011
+
+  static const int ANYONECANPAY = 0x80; // 1000 0000
+
+
+}